feat: 認証方法をRailsに対応
