+++ src/stores/useCookies.ts
@@ -1,9 +1,7 @@
 import { useCallback } from 'react'
 import { useCookies as useCookiesOriginal } from 'react-cookie'
-import { AuthInfo } from 'types/user/authInfo'
 
 type Cookies = {
-  authInfo: AuthInfo
   token: string
 }
