+++ src/types/user/authInfo.ts
@@ -1,19 +1,11 @@
-export type AuthInfo = {
-  token: string
-}
-
 export type LoginRequestParams = {
   email: string
   password: string
 }
 
-export type User = {
+export type SignUpRequestParams = {
+  username: string
   email: string
   password: string
-  createdAt: string
-  id: string
-  passwordDigest: string // ?
-  token: string
-  updatedAt: string
-  username: string
+  passwordConfirmation: string
 }
