--- src/types/user/form.ts
@@ -1,10 +1,0 @@
-// signUp, Login の型を定義
-
-export type SignUpFormParams = {
-  username: string
-  email: string
-  password: string
-  passwordConfirmation: string
-}
-
-export type SignUpRequest = Omit<SignUpFormParams, 'passwordConfirmation'>
