+++ src/types/user/user.ts
@@ -1,0 +1,11 @@
+// ユーザー情報更新
+export type User = {
+  email: string
+  password: string
+  createdAt: string
+  id: string
+  passwordDigest: string // いる？
+  token: string
+  updatedAt: string
+  username: string
+}
